Alterações de revisão do PR
