+++ src/@legacy/paginas/NAAPA/Encaminhamento/Cadastro/encaminhamentoNAAPA.js
@@ -46,7 +46,6 @@ const EncaminhamentoNAAPA = () => {
         <CadastroEncaminhamentoNAAPABotoesAcao
           mostrarBusca={mostrarBusca}
           setMostrarBusca={setMostrarBusca}
-          idEncaminhamento={encaminhamentoId}
         />
       </Cabecalho>
