+++ src/@legacy/servicos/Paginas/Gestao/NAAPA/ServicoNAAPA.js
@@ -44,10 +44,9 @@ class ServicoNAAPA {
         0}`
     );
 
-  imprimir = async idsSelecionados =>{
-     await api.post(`${URL_PADRAO}/imprimir-detalhado`, { encaminhamentoNaapaIds: idsSelecionados })
-      .catch(e => erros(e));
-  };
+  imprimir = idsSelecionados =>
+      api.post(`${URL_PADRAO}/imprimir-detalhado`, { encaminhamentoNaapaIds: idsSelecionados });
+  
 
   guardarSecaoEmEdicao = secaoId => {
     const { dispatch } = store;
